Fix #175: Channels now serialize and deserialize well
Issue #175 was that superclasses of Serializable classes that are not
themselves serializable will not have their fields serialized, and instead will
just have their zero-arg constructors called to create them. This was throwing
away the contents of the AbstractUpdatingInput object, specifically the
subscribers. Obviously, that doesn't work so well, so AbstractUpdatingInput is
now Serializable.

This also fixes the related issue that a wide variety of channel constructs
were not serializable. This is now easily fixed by simply making every single
channel serializable. Will it always serialize correctly? No. But this is a
step in the right direction.